requirements.txt
torch
torchvision
pandas
openpyxl
streamlit
Pillow
numpy
scikit-learn
matplotlib
tensorflow